requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.4
Babel==2.9.1
beautifulsoup4==4.9.3
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coverage==5.5
cryptography==3.4.7
cssselect==1.1.0
cssselect2==0.4.1
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-admin-honeypot==1.1.0
django-ajax-datatable==4.1.7
django-allauth==0.44.0
django-bootstrap-form==3.4
django-bootstrap4==3.0.1
django-crispy-forms==1.11.2
django-environ==0.4.5
django-extensions==3.1.2
django-filter==2.4.0
django-phonenumber-field==5.1.0
django-report-tools==0.2.2
django-sms==0.5.0
django-tables2==2.3.4
django-weasyprint==1.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
filelock==3.0.12
flake8==3.9.1
future==0.18.2
graphviz==0.16
gunicorn==20.1.0
html5lib==1.1
idna==2.10
lxml==4.6.3
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
pathlib==1.0.1
pdfrw==0.4
phonenumbers==8.12.22
Pillow==8.2.0
pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pydotplus==2.0.2
pyflakes==2.3.1
PyJWT==2.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2021.1
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
tox==3.23.0
urllib3==1.26.4
virtualenv==20.4.3
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
